tests: Move metric config fixture to conftest
